reorganise tstp/tptp statistics
